pmmgr: reduce unnecessary pmlogcheck burden
pmmgr runs pmlogcheck on archives that it's considering merging using
pmlogextract, because the latter can abort if any of its inputs are
corrupt, making no progress.  It turns out that this check should be
done after rather than before the granularity-mode logic, so that we
don't bother re-re-re-check very old or very new archives that we
aren't about to try to merge.  This speeds things up and has no QA
impact.